Moving helper methods in separate file
